+++ src/test/java/com/gmibank/pages/CommonWebElements.java
@@ -1,5 +1,6 @@
 package com.gmibank.pages;
 
+import com.gmibank.utilities.BrowserUtils;
 import com.gmibank.utilities.Driver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
@@ -55,5 +56,33 @@ public class CommonWebElements {
 
     @FindBys({@FindBy(xpath = "//table/tbody/tr")})
     public List<WebElement > allUsers;
+
+    @FindBy(id="login-item")
+    public WebElement signInLink;
+
+    @FindBy(name="username")
+    private WebElement usernameBox;
+
+    @FindBy(name="password")
+    private WebElement passwordBox;
+
+    @FindBy(xpath="//button/span[text()='Sign in']")
+    private WebElement signInBtn;
+
+    public void clickSignInLink(){
+        BrowserUtils.waitForClickablility(profileIcon,3).click();
+        BrowserUtils.waitForClickablility(signInLink,5).click();
+    }
+    public void enterUsername(String username){
+        BrowserUtils.waitForVisibility(usernameBox,3).sendKeys(username);
+    }
+
+    public void enterPassword(String password){
+        BrowserUtils.waitForVisibility(passwordBox,3).sendKeys(password);
+    }
+
+    public void clickSignInBtn(){
+        BrowserUtils.waitForClickablility(signInBtn,3).click();
+    }
 
 }
