+++ src/test/java/com/gmibank/runner/Runner.java
@@ -13,7 +13,7 @@ import org.junit.runner.RunWith;
         },
         features = "src/test/resources/features",
         glue = "com/gmibank/stepDefinitions",
-        tags = "@registrationPage",
+        tags = "",
         dryRun = false
 
 )
